✨ step3

+++ kvp-domain/src/main/java/com.kvp.domain/Introduce.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class Introduce {
     private String name;
-    private Long age;
+    private int age;
 
     public void addAge() {
         age += 10;
